src/test/javascript/spec/app/entities/jobs-manage/jobs-manage-update.component.spec.ts
/* tslint:disable max-line-length */
import { ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { Observable, of } from 'rxjs';

import { ELearningTestModule } from '../../../test.module';
import { JobsManageUpdateComponent } from 'app/entities/jobs-manage/jobs-manage-update.component';
import { JobsManageService } from 'app/entities/jobs-manage/jobs-manage.service';
import { JobsManage } from 'app/shared/model/jobs-manage.model';

describe('Component Tests', () => {
    describe('JobsManage Management Update Component', () => {
        let comp: JobsManageUpdateComponent;
        let fixture: ComponentFixture<JobsManageUpdateComponent>;
        let service: JobsManageService;

        beforeEach(() => {
            TestBed.configureTestingModule({
                imports: [ELearningTestModule],
                declarations: [JobsManageUpdateComponent]
            })
                .overrideTemplate(JobsManageUpdateComponent, '')
                .compileComponents();

            fixture = TestBed.createComponent(JobsManageUpdateComponent);
            comp = fixture.componentInstance;
            service = fixture.debugElement.injector.get(JobsManageService);
        });

        describe('save', () => {
            it(
                'Should call update service on save for existing entity',
                fakeAsync(() => {
                    // GIVEN
                    const entity = new JobsManage(123);
                    spyOn(service, 'update').and.returnValue(of(new HttpResponse({ body: entity })));
                    comp.jobs = entity;
                    // WHEN
                    comp.save();
                    tick(); // simulate async

                    // THEN
                    expect(service.update).toHaveBeenCalledWith(entity);
                    expect(comp.isSaving).toEqual(false);
                })
            );

            it(
                'Should call create service on save for new entity',
                fakeAsync(() => {
                    // GIVEN
                    const entity = new JobsManage();
                    spyOn(service, 'create').and.returnValue(of(new HttpResponse({ body: entity })));
                    comp.jobs = entity;
                    // WHEN
                    comp.save();
                    tick(); // simulate async

                    // THEN
                    expect(service.create).toHaveBeenCalledWith(entity);
                    expect(comp.isSaving).toEqual(false);
                })
            );
        });
    });
});
